+++ spinta/types/denorm/link.py
@@ -1,7 +1,7 @@
 from spinta import commands
 from spinta.components import Context, Property, Model
 from spinta.exceptions import NoRefPropertyForDenormProperty, ReferencedPropertyNotFound
-from spinta.types.datatype import Denorm, Ref, Object
+from spinta.types.datatype import Denorm, Ref, Object, Array
 from spinta.types.helpers import set_dtype_backend
 
 
@@ -16,6 +16,7 @@ def link(context: Context, dtype: Denorm) -> None:
     )
 
 
+# TODO: Add better support for denorm when nested (with object type, etc.)
 def _get_denorm_prop(
     name: str,
     prop: Property,
@@ -24,9 +25,15 @@ def _get_denorm_prop(
     models = model.manifest.models
     name_parts = name.split('.', 1)
     name = name_parts[0]
-    properties = model.properties
+    properties = prop.parent.dtype.model.properties if isinstance(prop.parent.dtype, Ref) else prop.parent.model.properties
+    model = models[prop.parent.dtype.model.name] if isinstance(prop.parent.dtype, Ref) else model
     if len(name_parts) > 1:
-        if name not in properties or not isinstance(properties[name].dtype, (Ref, Object)):
+        ref_prop = properties[name]
+        while isinstance(ref_prop.dtype, Array):
+            ref_prop = ref_prop.dtype.items
+
+        model = models[ref_prop.dtype.model.name] if isinstance(ref_prop.dtype, Ref) else model
+        if name not in properties or not isinstance(ref_prop.dtype, (Ref, Object)):
             if prop.model == model:
                 raise NoRefPropertyForDenormProperty(
                     prop,
@@ -39,8 +46,6 @@ def _get_denorm_prop(
                     ref={'property': name, 'model': model.name},
                 )
         else:
-            ref_prop = properties[name]
-            model = models[ref_prop.dtype.model.name]
             denorm_prop = _get_denorm_prop(name_parts[1], prop, model)
     else:
         if name not in properties: