+++ tests/manifests/test_tabular.py
@@ -1,6 +1,7 @@
 import pytest
 
-from spinta.exceptions import InvalidManifestFile, NoRefPropertyForDenormProperty, ReferencedPropertyNotFound
+from spinta.exceptions import InvalidManifestFile, ModelReferenceNotFound, ReferencedPropertyNotFound, \
+    PartialTypeNotFound, DataTypeCannotBeUsedForNesting, NestedDataTypeMissmatch
 from spinta.testing.tabular import create_tabular_manifest
 from spinta.testing.manifest import load_manifest
 from spinta.manifests.tabular.helpers import TabularManifestError
@@ -429,7 +430,7 @@ def test_with_denormalized_data(tmp_path, rc):
     
     
 def test_with_denormalized_data_ref_error(tmp_path, rc):
-    with pytest.raises(NoRefPropertyForDenormProperty) as e:
+    with pytest.raises(PartialTypeNotFound) as e:
         check(tmp_path, rc, '''
         d | r | b | m | property               | type   | ref       | access
         example                                |        |           |
@@ -441,10 +442,6 @@ def test_with_denormalized_data_ref_error(tmp_path, rc):
           |   |   |   | name                   | string |           | open
           |   |   |   | country.name           |        |           | open
         ''')
-    assert e.value.message == (
-        "Property 'country' with type 'ref' or 'object' must be defined "
-        "before defining property 'country.name'."
-    )
 
 
 def test_with_denormalized_data_undefined_error(tmp_path, rc):
@@ -775,3 +772,207 @@ def test_prop_array_customize_type(tmp_path, rc):
           |   |   |   | languages   | array   |          | open   | Array of languages
           |   |   |   | languages[] | string  |          | open   | Correction
     ''')
+
+
+def test_prop_multi_array(tmp_path, rc):
+    check(tmp_path, rc, '''
+        d | r | b | m | property        | type    | ref      | access | title
+        example                         |         |          |        |
+                                        |         |          |        |
+          |   |   | Country             |         |          |        |
+          |   |   |   | name            | string  |          | open   |
+          |   |   |   | languages[][][] | string  |          | open   | Correction
+    ''')
+
+
+def test_prop_multi_array_specific(tmp_path, rc):
+    check(tmp_path, rc, '''
+        d | r | b | m | property        | type    | ref      | access | title
+        example                         |         |          |        |
+                                        |         |          |        |
+          |   |   | Country             |         |          |        |
+          |   |   |   | name            | string  |          | open   |
+          |   |   |   | languages       | array   |          | open   | Correction T0
+          |   |   |   | languages[]     | array   |          | open   | Correction T1
+          |   |   |   | languages[][]   | array   |          | open   | Correction T2
+          |   |   |   | languages[][][] | string  |          | open   | Correction T3
+    ''')
+
+
+def test_prop_nested_denorm(tmp_path, rc):
+    check(tmp_path, rc, '''
+        d | r | b | m | property        | type    | ref      | access | title
+        example                         |         |          |        |
+                                        |         |          |        |
+          |   |   | Language            |         |          |        |
+          |   |   |   | dialect         | string  |          | open   |
+                                        |         |          |        |
+          |   |   | Country             |         |          |        |
+          |   |   |   | name            | string  |          | open   |
+          |   |   |   | langs[]         | ref     | Language | open   |
+          |   |   |   | langs[].dialect |         |          | open   | Denorm
+    ''')
+
+
+def test_prop_multi_nested_denorm(tmp_path, rc):
+    check(tmp_path, rc, '''
+        d | r | b | m | property          | type    | ref      | access | title
+        example                           |         |          |        |
+                                          |         |          |        |
+          |   |   | Language              |         |          |        |
+          |   |   |   | dialect           | string  |          | open   |
+                                          |         |          |        |
+          |   |   | Country               |         |          |        |
+          |   |   |   | name              | string  |          | open   |
+          |   |   |   | langs             | array   |          | open   |
+          |   |   |   | langs[]           | array   |          | open   |
+          |   |   |   | langs[][]         | ref     | Language | open   |
+          |   |   |   | langs[][].dialect |         |          | open   |
+    ''')
+
+
+def test_prop_multi_nested_error_partial(tmp_path, rc):
+    with pytest.raises(PartialTypeNotFound) as e:
+        check(tmp_path, rc, '''
+            d | r | b | m | property          | type    | ref      | access | title
+            example                           |         |          |        |
+                                              |         |          |        |
+              |   |   | Language              |         |          |        |
+              |   |   |   | dialect           | string  |          | open   |
+                                              |         |          |        |
+              |   |   | Country               |         |          |        |
+              |   |   |   | name              | string  |          | open   |
+              |   |   |   | langs             | array   |          | open   |
+              |   |   |   | langs[][].dialect |         |          | open   |
+        ''')
+
+
+def test_prop_multi_nested_multi_models(tmp_path, rc):
+    check(tmp_path, rc, '''
+        d | r | b | m | property               | type    | ref       | access | title
+        example                                |         |           |        |
+                                               |         |           |        |
+          |   |   | Continent                  |         | id        |        |
+          |   |   |   | id                     | integer |           | open   |
+          |   |   |   | name                   | string  |           | open   |
+                                               |         |           |        |
+          |   |   | Country                    |         | id        |        |
+          |   |   |   | id                     | integer |           | open   |
+          |   |   |   | name                   | string  |           | open   |
+          |   |   |   | continent              | ref     | Continent | open   |
+                                               |         |           |        |
+          |   |   | City                       |         | id        |        |
+          |   |   |   | id                     | integer |           | open   |
+          |   |   |   | country                | ref     | Country   | open   |
+          |   |   |   | country.code           | string  |           | open   |
+          |   |   |   | country.name           |         |           | open   |
+          |   |   |   | country.continent.code | string  |           | open   |
+          |   |   |   | country.continent.name |         |           | open   |
+    ''')
+
+
+def test_prop_multi_nested(tmp_path, rc):
+    check(tmp_path, rc, '''
+            d | r | b | m | property                       | type    | ref      | access | title
+            example                                        |         |          |        |
+                                                           |         |          |        |
+              |   |   | Language                           |         |          |        |
+              |   |   |   | dialect                        | string  |          | open   |
+              |   |   |   | meta                           | object  |          | open   |
+              |   |   |   | meta.version                   | integer |          | open   |
+                                                           |         |          |        |
+              |   |   | Country                            |         |          |        |
+              |   |   |   | name                           | string  |          | open   |
+              |   |   |   | meta                           | object  |          | open   |
+              |   |   |   | meta.version                   | integer |          | open   |
+              |   |   |   | meta.old                       | object  |          | open   |
+              |   |   |   | meta.old.language              | ref     | Language | open   |
+              |   |   |   | meta.old.language.dialect      |         |          | open   |
+              |   |   |   | meta.old.language.meta.version | integer |          | open   |
+              |   |   |   | meta.langs                     | array   |          | open   |
+              |   |   |   | meta.langs[]                   | array   |          | open   |
+              |   |   |   | meta.langs[][]                 | ref     | Language | open   |
+              |   |   |   | meta.langs[][].dialect         |         |          | open   |
+        ''')
+
+
+def test_multi_nested_incorrect(tmp_path, rc):
+    with pytest.raises(DataTypeCannotBeUsedForNesting) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta.version                   | string  |          | open   |
+                  |   |   |   | meta                           | integer |          | open   |
+            ''')
+
+
+def test_multi_nested_incorrect_reversed_order(tmp_path, rc):
+    with pytest.raises(DataTypeCannotBeUsedForNesting) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta                           | integer |          | open   |
+                  |   |   |   | meta.version                   | string  |          | open   |
+            ''')
+
+
+def test_multi_nested_incorrect_deep(tmp_path, rc):
+    with pytest.raises(DataTypeCannotBeUsedForNesting) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta.version.id                | integer |          | open   |
+                  |   |   |   | meta.version                   | string  |          | open   |
+                  |   |   |   | meta                           | object  |          | open   |
+            ''')
+
+
+def test_multi_nested_incorrect_with_array(tmp_path, rc):
+    with pytest.raises(DataTypeCannotBeUsedForNesting) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta.version[].id              | integer |          | open   |
+                  |   |   |   | meta.version[]                 | string  |          | open   |
+                  |   |   |   | meta                           | object  |          | open   |
+            ''')
+
+
+def test_multi_nested_type_missmatch_with_array(tmp_path, rc):
+    with pytest.raises(NestedDataTypeMissmatch) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta.version.id                | integer |          | open   |
+                  |   |   |   | meta.version[]                 | string  |          | open   |
+                  |   |   |   | meta                           | object  |          | open   |
+            ''')
+
+
+def test_multi_nested_type_missmatch_with_partial(tmp_path, rc):
+    with pytest.raises(NestedDataTypeMissmatch) as e:
+        check(tmp_path, rc, '''
+                d | r | b | m | property                       | type    | ref      | access | title
+                example                                        |         |          |        |
+                                                               |         |          |        |
+                  |   |   | Language                           |         |          |        |
+                  |   |   |   | dialect                        | string  |          | open   |
+                  |   |   |   | meta.version[]                 | string  |          | open   |
+                  |   |   |   | meta.version.id                | integer |          | open   |
+                  |   |   |   | meta                           | object  |          | open   |
+            ''')